Deleted functions that were present in several files and moved those functions into the auxiliary_functions.py script in the utils directory. The auxiliary_functions will be imported so that the functions can be used. The NSStats process now works from the CMD.
