+++ CHEWBBACA/CHEWBBACA_NS/stats_requests.py
@@ -1,21 +1,23 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Created on Fri Oct 18 15:37:16 2019
+AUTHORS
 
-@author: rfm
+    Mickael Silva
+    github: @
+
+    Pedro Cerqueira
+    github: @pedrorvc
+
+    Rafael Mamede
+    github: @rfm-targa
 
 DESCRIPTION
-
-    Script used by users to get useful stats:
-        -
 
 """
 
 
-import sys
 import argparse
-import requests
 import concurrent.futures
 from getpass import getpass
 
@@ -23,24 +25,10 @@ from SPARQLWrapper import SPARQLWrapper
 
 from utils import auxiliary_functions as aux
 
-# way of importing scripts or modules that are not in the current directory
-#sys.path.insert(1, '/home/rfm/Desktop/rfm/Lab_Software/Chewie_NS/NS_tests/ns_security_backup_20191009')
-#
-#
-#virtuoso_server = SPARQLWrapper(app.config['LOCAL_SPARQL'])
-#virtuoso_server = SPARQLWrapper('http://localhost:8890/sparql')
-#url_send_local_virtuoso = 'http://localhost:8890/DAV/test_folder/data'
 
-
-def simple_request(base_url, headers_get, endpoint_list):
-    """
-    """
-
-    # unpack list of sequential endpoints and pass to create URI
-    url = aux.make_url(base_url, *endpoint_list)
-    res = requests.get(url, headers=headers_get, timeout=30)
-
-    return res
+# virtuoso_server = SPARQLWrapper(app.config['LOCAL_SPARQL'])
+# virtuoso_server = SPARQLWrapper('http://localhost:8890/sparql')
+# url_send_local_virtuoso = 'http://localhost:8890/DAV/test_folder/data'
 
 
 def species_stats(base_url, headers_get, endpoint_list):
@@ -48,17 +36,17 @@ def species_stats(base_url, headers_get, endpoint_list):
     """
 
     # unpack list of sequential endpoints and pass to create URI
-    species_namids = species_list(base_url, headers_get, endpoint_list)
+    species_namids = aux.species_list(base_url, headers_get, endpoint_list)
 
     species_stats = []
     species_stats.append('{:<30}  {:^10}  {:^10}  {:^10}'.format('Species',
-                                                                         'id',
-                                                                         '#schemas',
-                                                                         '#loci'))
-                                                                         #'#alleles'))
+                                                                 'id',
+                                                                 '#schemas',
+                                                                 '#loci'))
+                                                                 # 'alleles'))
     for sp in species_namids:
-        species = sp[0]
-        species_id = sp[1]
+        species = sp
+        species_id = species_namids[sp]
 
         # get number of schemas
         species_schemas = count_schemas(base_url,
@@ -76,7 +64,7 @@ def species_stats(base_url, headers_get, endpoint_list):
 #            species_alleles += count_loci_alleles(base_url,
 #                                                  headers_get,
 #                                                  ['loci', locus, 'alleles'])
-        
+
         # get total number of isolates
 
 
@@ -90,29 +78,12 @@ def species_stats(base_url, headers_get, endpoint_list):
     return species_stats
 
 
-def species_list(base_url, headers_get, endpoint_list):
-    """
-    """
-
-    res = simple_request(base_url, headers_get, endpoint_list)
-    res = res.json()
-    species_lst = []
-    for sp in res:
-        species = sp['name']['value']
-        species_url = sp['species']['value']
-        species_id = species_url.split('/')[-1]
-
-        species_lst.append((species, species_id))
-
-    return species_lst
-
-
 def count_loci(base_url, headers_get, endpoint_list):
     """
     """
 
     # unpack list of sequential endpoints and pass to create URI
-    res = simple_request(base_url, headers_get, endpoint_list)
+    res = aux.simple_get_request(base_url, headers_get, endpoint_list)
     res = res.json()
 
     loci_ids = []
@@ -135,7 +106,7 @@ def count_schemas(base_url, headers_get, endpoint_list):
     """
 
     # unpack list of sequential endpoints and pass to create URI
-    res = simple_request(base_url, headers_get, endpoint_list)
+    res = aux.simple_get_request(base_url, headers_get, endpoint_list)
     res = res.json()
 
     if 'message' not in res:
@@ -151,7 +122,7 @@ def schemas_info(base_url, headers_get, endpoint_list):
     """
 
     # unpack list of sequential endpoints and pass to create URI
-    res = simple_request(base_url, headers_get, endpoint_list)
+    res = aux.simple_get_request(base_url, headers_get, endpoint_list)
     res = res.json()
 
     sc_info = {}
@@ -172,7 +143,7 @@ def count_loci_alleles(loci_info):
     endpoint_list = loci_info[2]
 
     # unpack list of sequential endpoints and pass to create URI
-    res = simple_request(base_url, headers_get, endpoint_list)
+    res = aux.simple_get_request(base_url, headers_get, endpoint_list)
     res = res.json()
 
     total_alleles = len(res)
@@ -187,10 +158,10 @@ def species_schemas_stats(base_url, headers_get, endpoint_list, species_id, work
     given = species_id
 
     # unpack list of sequential endpoints and pass to create URI
-    species_namids = species_list(base_url, headers_get, endpoint_list)
+    species_namids = aux.species_list(base_url, headers_get, endpoint_list)
 
-    species_names = {t[0]:t[1] for t in species_namids}
-    species_ids = {t[1]:t[0] for t in species_namids}
+    species_names = {k: v for k, v in species_namids.items()}
+    species_ids = {v: k for k, v in species_namids.items()}
 
     sp = species_ids.get(species_id, 0)
     if sp == 0:
@@ -199,12 +170,20 @@ def species_schemas_stats(base_url, headers_get, endpoint_list, species_id, work
         if species_id == 0:
             return 'There is no taxon name or identifier "{0}" in server.'.format(given)
 
+    print('Fetching data about {0} available schemas...\n'.format(sp))
+
     schema_stats = []
-    schema_stats.append('{0} available schemas\n'.format(sp))
-    schema_stats.append('{:<10}  {:^10}  {:^20}'.format('Schema_id',
-                                                                '#loci',
-                                                                '#alleles',
-                                                                'description'))
+    table_title = '{0} available schemas\n'.format(sp)
+    print(table_title)
+    schema_stats.append(table_title)
+    print('-'*50)
+    header = '{:<10}  {:^10}  {:^10}  {:^20}'.format('Schema_id',
+                                                     '#loci',
+                                                     '#alleles',
+                                                     'description')
+    print(header)
+    schema_stats.append(header)
+    print('-'*50)
 
     # get schemas for selected species
     schemas = schemas_info(base_url, headers_get, ['species', species_id, 'schemas'])
@@ -220,16 +199,18 @@ def species_schemas_stats(base_url, headers_get, endpoint_list, species_id, work
         # get number of alleles
         schema_alleles = 0
         schema_loci_ids = [[base_url, headers_get, ['loci', locus, 'alleles']] for locus in schema_loci_ids]
-        #print(schema_loci_ids)
         with concurrent.futures.ThreadPoolExecutor(max_workers=workers) as executor:
             for res in executor.map(count_loci_alleles, schema_loci_ids):
                 schema_alleles += res
-            print(schema_alleles)
 
-        schema_stats.append('{:^10}  {:^10}  {:^10}  {:^20}'.format(sid,
-                                                                    schema_total_loci,
-                                                                    schema_alleles,
-                                                                    sdesc))
+        schema_line = '{:^10}  {:^10}  {:^10}  {:^20}'.format(sid,
+                                                              schema_total_loci,
+                                                              schema_alleles,
+                                                              sdesc)
+        print(schema_line)
+        schema_stats.append(schema_line)
+
+    print('-'*50)
 
     return schema_stats
 
@@ -253,7 +234,7 @@ def parse_arguments():
 
     return [args.stats_mode, args.base_url, args.taxon]
 
-# add multithreading to GET number of alleles
+
 def main(mode, base_url, taxid):
     """
     """
@@ -267,7 +248,7 @@ def main(mode, base_url, taxid):
     # get token
     token = aux.login_user_to_NS(base_url, user, password)
 
-    headers_get = {'X-API-KEY': token,
+    headers_get = {'Authorization': token,
                    'accept': 'application/json'}
 
     if mode == 'species_list':
@@ -288,8 +269,9 @@ def main(mode, base_url, taxid):
                                           endpoint_list, species_id, 30)
 
             # print stats
-            stats_text = '\n'.join(stats)
-            print(stats_text)
+#            stats_text = '\n'.join(stats)
+#            print(stats_text)
+            print('\n')
         else:
             print('\nPlease provide a valid taxon name or identifier '
                   'to get the list of available schemas.\n')
